+++ packages/iabtcf_consent_info_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [next]
+
+- **fix**: clarify semantics of stream returned by `rawConsentInfo`
+
 # [1.0.0] - 2020-12-24
 
 - Initial release